utils/libs/config/__init__.py
import os
from dotenv import load_dotenv

load_dotenv()

MONGODB_URI = os.getenv("MONGODB_URI")
MONGODB_DB = os.getenv("MONGODB_DB")
MONGODB_COLLECTION = os.getenv("MONGODB_COLLECTION_USERS")



PORT = os.getenv("PORT")
HOST = os.getenv("HOST")