requirement.txt
fastapi = 0.115.12
uvicorn = 0.34.2
python-dotenv=1.1.0
pymongo = 4.13.0

